refactor(core.identity)!: extract identity stores and simplify temp api
BREAKING CHANGE: Removed temporary identity-specific methods from IdentityService; use deleteIdentity and createTemporaryIdentity.
