+++ kura/org.eclipse.kura.core.identity/src/main/java/org/eclipse/kura/core/identity/PasswordHasher.java
@@ -1,0 +1,23 @@
+/*******************************************************************************
+ * Copyright (c) 2026 Eurotech and/or its affiliates and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *  Eurotech
+ *******************************************************************************/
+// Content with portions generated by generative AI platform
+
+package org.eclipse.kura.core.identity;
+
+import org.eclipse.kura.KuraException;
+import org.eclipse.kura.identity.PasswordHash;
+
+public interface PasswordHasher {
+
+    PasswordHash hash(char[] password) throws KuraException;
+}
