+++ kura/org.eclipse.kura.core.identity/src/main/java/org/eclipse/kura/core/identity/store/IdentityStore.java
@@ -1,0 +1,38 @@
+/*******************************************************************************
+ * Copyright (c) 2026 Eurotech and/or its affiliates and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *  Eurotech
+ *******************************************************************************/
+// Content with portions generated by generative AI platform
+
+package org.eclipse.kura.core.identity.store;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
+
+import org.eclipse.kura.KuraException;
+import org.eclipse.kura.identity.IdentityConfiguration;
+import org.eclipse.kura.identity.IdentityConfigurationComponent;
+
+interface IdentityStore {
+
+    boolean exists(String identityName) throws KuraException;
+
+    Optional<IdentityConfiguration> getIdentityConfiguration(String identityName,
+            Set<Class<? extends IdentityConfigurationComponent>> componentsToReturn) throws KuraException;
+
+    List<IdentityConfiguration> getIdentitiesConfiguration(
+            Set<Class<? extends IdentityConfigurationComponent>> componentsToReturn) throws KuraException;
+
+    void updateIdentityConfiguration(IdentityConfiguration identityConfiguration) throws KuraException;
+
+    boolean deleteIdentity(String identityName) throws KuraException;
+}
