+++ kura/org.eclipse.kura.core.identity/src/main/java/org/eclipse/kura/core/identity/store/TemporaryIdentityStore.java
@@ -12,7 +12,7 @@
  *******************************************************************************/
 // Content with portions generated by generative AI platform
 
-package org.eclipse.kura.core.identity;
+package org.eclipse.kura.core.identity.store;
 
 import java.time.Duration;
 import java.time.Instant;
