+++ kura/org.eclipse.kura.core.identity/src/main/java/org/eclipse/kura/core/identity/store/TemporaryIdentityStoreAdapter.java
@@ -1,0 +1,186 @@
+/*******************************************************************************
+ * Copyright (c) 2026 Eurotech and/or its affiliates and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *  Eurotech
+ *******************************************************************************/
+// Content with portions generated by generative AI platform
+
+package org.eclipse.kura.core.identity.store;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.time.Duration;
+
+import org.eclipse.kura.KuraErrorCode;
+import org.eclipse.kura.KuraException;
+import org.eclipse.kura.core.identity.PasswordHasher;
+import org.eclipse.kura.core.identity.ValidationUtil;
+import org.eclipse.kura.identity.IdentityConfiguration;
+import org.eclipse.kura.identity.IdentityConfigurationComponent;
+import org.eclipse.kura.identity.PasswordConfiguration;
+import org.eclipse.kura.identity.PasswordHash;
+import org.eclipse.kura.identity.PasswordStrengthVerificationService;
+
+class TemporaryIdentityStoreAdapter implements IdentityStore {
+
+    private final TemporaryIdentityStore temporaryStore;
+    private final PasswordStrengthVerificationService passwordStrengthVerificationService;
+    private final PasswordHasher passwordHasher;
+
+    TemporaryIdentityStoreAdapter(final TemporaryIdentityStore temporaryStore,
+            final PasswordStrengthVerificationService passwordStrengthVerificationService,
+            final PasswordHasher passwordHasher) {
+        this.temporaryStore = temporaryStore;
+        this.passwordStrengthVerificationService = passwordStrengthVerificationService;
+        this.passwordHasher = passwordHasher;
+    }
+
+    @Override
+    public boolean exists(final String identityName) {
+        return this.temporaryStore.exists(identityName);
+    }
+
+    @Override
+    public Optional<IdentityConfiguration> getIdentityConfiguration(final String identityName,
+            final Set<Class<? extends IdentityConfigurationComponent>> componentsToReturn) {
+        return this.temporaryStore.getIdentity(identityName)
+                .map(config -> filterIdentityConfiguration(config, componentsToReturn));
+    }
+
+    @Override
+    public List<IdentityConfiguration> getIdentitiesConfiguration(
+            final Set<Class<? extends IdentityConfigurationComponent>> componentsToReturn) {
+        final List<IdentityConfiguration> result = new ArrayList<>();
+
+        for (final String identityName : this.temporaryStore.getAllIdentityNames()) {
+            this.temporaryStore.getIdentity(identityName)
+                    .map(config -> filterIdentityConfiguration(config, componentsToReturn))
+                    .ifPresent(result::add);
+        }
+
+        return result;
+    }
+
+    @Override
+    public void updateIdentityConfiguration(final IdentityConfiguration identityConfiguration) throws KuraException {
+        final IdentityConfiguration existingConfiguration = this.temporaryStore
+                .getIdentity(identityConfiguration.getName())
+                .orElseGet(() -> new IdentityConfiguration(identityConfiguration.getName(), Collections.emptyList()));
+
+        final IdentityConfiguration processedConfiguration = processConfigurationForTemporaryStorage(
+                identityConfiguration, Optional.of(existingConfiguration));
+
+        final Map<Class<? extends IdentityConfigurationComponent>, IdentityConfigurationComponent> merged =
+                new HashMap<>();
+        for (final IdentityConfigurationComponent component : existingConfiguration.getComponents()) {
+            merged.put(component.getClass(), component);
+        }
+        for (final IdentityConfigurationComponent component : processedConfiguration.getComponents()) {
+            merged.put(component.getClass(), component);
+        }
+
+        final List<IdentityConfigurationComponent> mergedComponents = new ArrayList<>(merged.values());
+        final IdentityConfiguration mergedConfiguration = new IdentityConfiguration(identityConfiguration.getName(),
+                mergedComponents);
+
+        if (!this.temporaryStore.updateIdentity(identityConfiguration.getName(), mergedConfiguration)) {
+            throw new KuraException(KuraErrorCode.INVALID_PARAMETER, "Temporary identity does not exist");
+        }
+    }
+
+    @Override
+    public boolean deleteIdentity(final String identityName) {
+        return this.temporaryStore.deleteIdentity(identityName);
+    }
+
+    void createIdentity(final IdentityConfiguration configuration, final Duration lifetime) throws KuraException {
+        final IdentityConfiguration processedConfiguration = processConfigurationForTemporaryStorage(configuration,
+                Optional.empty());
+        this.temporaryStore.createIdentity(configuration.getName(), processedConfiguration, lifetime);
+    }
+
+    private IdentityConfiguration filterIdentityConfiguration(final IdentityConfiguration configuration,
+            final Set<Class<? extends IdentityConfigurationComponent>> componentsToReturn) {
+        if (componentsToReturn.isEmpty()) {
+            return new IdentityConfiguration(configuration.getName(), Collections.emptyList());
+        }
+
+        final List<IdentityConfigurationComponent> filteredComponents = configuration.getComponents().stream()
+                .filter(component -> componentsToReturn.stream().anyMatch(clazz -> clazz.isInstance(component)))
+                .toList();
+
+        return new IdentityConfiguration(configuration.getName(), filteredComponents);
+    }
+
+    private PasswordConfiguration processPasswordForStorage(final String identityName,
+            final PasswordConfiguration passwordConfiguration, final Optional<PasswordHash> existingPasswordHash)
+            throws KuraException {
+
+        if (!passwordConfiguration.isPasswordAuthEnabled()) {
+            return new PasswordConfiguration(
+                    passwordConfiguration.isPasswordChangeNeeded(),
+                    false,
+                    Optional.empty(),
+                    Optional.empty());
+        }
+
+        final Optional<char[]> newPassword = passwordConfiguration.getNewPassword();
+
+        if (newPassword.isPresent()) {
+            ValidationUtil.validateNewPassword(identityName, newPassword.get(), this.passwordStrengthVerificationService);
+
+            final PasswordHash hash = this.passwordHasher.hash(newPassword.get());
+
+            return new PasswordConfiguration(
+                    passwordConfiguration.isPasswordChangeNeeded(),
+                    true,
+                    Optional.empty(),
+                    Optional.of(hash));
+        }
+
+        if (!existingPasswordHash.isPresent()) {
+            throw new KuraException(KuraErrorCode.INVALID_PARAMETER,
+                    "Password authentication is enabled but no password has been provided");
+        }
+
+        return new PasswordConfiguration(
+                passwordConfiguration.isPasswordChangeNeeded(),
+                true,
+                Optional.empty(),
+                existingPasswordHash);
+    }
+
+    private IdentityConfiguration processConfigurationForTemporaryStorage(
+            final IdentityConfiguration configuration, final Optional<IdentityConfiguration> existingConfiguration)
+            throws KuraException {
+
+        final String identityName = configuration.getName();
+        final Optional<PasswordHash> existingPasswordHash = existingConfiguration
+                .flatMap(config -> config.getComponent(PasswordConfiguration.class))
+                .flatMap(PasswordConfiguration::getPasswordHash);
+        final List<IdentityConfigurationComponent> processedComponents = new ArrayList<>();
+
+        for (final IdentityConfigurationComponent component : configuration.getComponents()) {
+            if (component instanceof PasswordConfiguration) {
+                processedComponents.add(
+                        processPasswordForStorage(identityName, (PasswordConfiguration) component, existingPasswordHash));
+            } else {
+                processedComponents.add(component);
+            }
+        }
+
+        return new IdentityConfiguration(identityName, processedComponents);
+    }
+}
