+++ kura/org.eclipse.kura.core.identity/src/main/java/org/eclipse/kura/core/identity/store/UserAdminIdentityStore.java
@@ -1,0 +1,283 @@
+/*******************************************************************************
+ * Copyright (c) 2026 Eurotech and/or its affiliates and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *  Eurotech
+ *******************************************************************************/
+// Content with portions generated by generative AI platform
+
+package org.eclipse.kura.core.identity.store;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Dictionary;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
+
+import org.eclipse.kura.KuraErrorCode;
+import org.eclipse.kura.KuraException;
+import org.eclipse.kura.configuration.ComponentConfiguration;
+import org.eclipse.kura.core.identity.IdentityServiceImpl;
+import org.eclipse.kura.core.identity.PasswordHasher;
+import org.eclipse.kura.identity.AdditionalConfigurations;
+import org.eclipse.kura.identity.AssignedPermissions;
+import org.eclipse.kura.identity.IdentityConfiguration;
+import org.eclipse.kura.identity.IdentityConfigurationComponent;
+import org.eclipse.kura.identity.PasswordConfiguration;
+import org.eclipse.kura.identity.Permission;
+import org.eclipse.kura.identity.PasswordHashImpl;
+import org.eclipse.kura.identity.configuration.extension.IdentityConfigurationExtension;
+import org.eclipse.kura.util.useradmin.UserAdminHelper;
+import org.osgi.service.useradmin.Role;
+import org.osgi.service.useradmin.User;
+import org.slf4j.Logger;
+
+class UserAdminIdentityStore implements IdentityStore {
+
+    private final UserAdminHelper userAdminHelper;
+    private final Map<String, IdentityConfigurationExtension> extensions;
+    private final Logger logger;
+    private final PasswordHasher passwordHasher;
+
+    UserAdminIdentityStore(final UserAdminHelper userAdminHelper,
+            final Map<String, IdentityConfigurationExtension> extensions,
+            final Logger logger,
+            final PasswordHasher passwordHasher) {
+        this.userAdminHelper = userAdminHelper;
+        this.extensions = extensions;
+        this.logger = logger;
+        this.passwordHasher = passwordHasher;
+    }
+
+    @Override
+    public boolean exists(final String identityName) {
+        return this.userAdminHelper.getUser(identityName).isPresent();
+    }
+
+    @Override
+    public Optional<IdentityConfiguration> getIdentityConfiguration(final String identityName,
+            final Set<Class<? extends IdentityConfigurationComponent>> componentsToReturn) {
+        return this.userAdminHelper.getUser(identityName)
+                .map(user -> buildIdentity(identityName, user, componentsToReturn));
+    }
+
+    @Override
+    public List<IdentityConfiguration> getIdentitiesConfiguration(
+            final Set<Class<? extends IdentityConfigurationComponent>> componentsToReturn) {
+        final List<IdentityConfiguration> result = new ArrayList<>();
+        this.userAdminHelper.foreachUser((name, user) -> result.add(buildIdentity(name, user, componentsToReturn)));
+        return result;
+    }
+
+    @Override
+    public void updateIdentityConfiguration(final IdentityConfiguration identityConfiguration) throws KuraException {
+        final User user = this.userAdminHelper.getUser(identityConfiguration.getName())
+                .orElseThrow(() -> new KuraException(KuraErrorCode.INVALID_PARAMETER, "Identity does not exist"));
+        updateIdentityConfigurationInternal(user, identityConfiguration);
+    }
+
+    @Override
+    public boolean deleteIdentity(final String identityName) throws KuraException {
+        if (!this.userAdminHelper.getUser(identityName).isPresent()) {
+            return false;
+        }
+
+        this.userAdminHelper.deleteUser(identityName);
+        return true;
+    }
+
+    private void updateIdentityConfigurationInternal(final User user, final IdentityConfiguration identity)
+            throws KuraException {
+
+        final String identityName = identity.getName();
+        final Optional<PasswordConfiguration> passwordData = identity.getComponent(PasswordConfiguration.class);
+
+        if (passwordData.isPresent()) {
+            updatePassword(identityName, passwordData.get(), user);
+        }
+
+        final Optional<AssignedPermissions> permissions = identity.getComponent(AssignedPermissions.class);
+
+        if (permissions.isPresent()) {
+            updateAssignedPermissions(identityName, permissions.get(), user);
+        }
+
+        final Optional<AdditionalConfigurations> additionalConfigurations = identity
+                .getComponent(AdditionalConfigurations.class);
+
+        if (additionalConfigurations.isPresent()) {
+            updateAdditionalConfigurations(identity.getName(), additionalConfigurations.get());
+        }
+    }
+
+    private IdentityConfiguration buildIdentity(final String name, final User user,
+            final Set<Class<? extends IdentityConfigurationComponent>> componentsToReturn) {
+
+        final List<IdentityConfigurationComponent> components = new ArrayList<>();
+
+        if (componentsToReturn.contains(PasswordConfiguration.class)) {
+            components.add(getPasswordData(user));
+        }
+
+        if (componentsToReturn.contains(AssignedPermissions.class)) {
+            final Set<Permission> permissions = this.userAdminHelper.getIdentityPermissions(name).stream()
+                    .map(Permission::new).collect(Collectors.toSet());
+            components.add(new AssignedPermissions(permissions));
+        }
+
+        if (componentsToReturn.contains(AdditionalConfigurations.class)) {
+            components.add(getAdditionalConfigurations(name));
+        }
+
+        return new IdentityConfiguration(name, components);
+    }
+
+    private AdditionalConfigurations getAdditionalConfigurations(final String name) {
+        final List<ComponentConfiguration> additionalConfigurations = new ArrayList<>();
+
+        for (final IdentityConfigurationExtension extension : this.extensions.values()) {
+            try {
+                extension.getConfiguration(name).ifPresent(additionalConfigurations::add);
+            } catch (final Exception e) {
+                this.logger.warn("Failed to get identity additional configuration from extension", e);
+            }
+        }
+
+        return new AdditionalConfigurations(additionalConfigurations);
+    }
+
+    private PasswordConfiguration getPasswordData(final User user) {
+        final Optional<String> passwordHash = Optional.ofNullable(user.getCredentials()
+                .get(IdentityServiceImpl.PASSWORD_PROPERTY))
+                .filter(String.class::isInstance).map(String.class::cast);
+
+        final boolean isPasswordChangeNeeded = Objects.equals("true",
+                user.getProperties().get(IdentityServiceImpl.KURA_NEED_PASSWORD_CHANGE));
+
+        return new PasswordConfiguration(isPasswordChangeNeeded, passwordHash.isPresent(), Optional.empty(),
+                passwordHash.map(PasswordHashImpl::new));
+    }
+
+    private void updateAssignedPermissions(final String identityName, final AssignedPermissions assignedPermissions,
+            final User user) {
+        this.userAdminHelper.foreachPermission((name, group) -> {
+            final Permission permission = new Permission(name);
+            final List<Role> members = Optional.ofNullable(group.getMembers()).map(Arrays::asList)
+                    .orElse(Collections.emptyList());
+
+            if (assignedPermissions.getPermissions().contains(permission) && !members.contains(user)) {
+                IdentityServiceImpl.audit(() -> group.addMember(user),
+                        "Add permission " + permission.getName() + " to identity " + identityName);
+            } else if (!assignedPermissions.getPermissions().contains(permission) && members.contains(user)) {
+                IdentityServiceImpl.audit(() -> group.removeMember(user),
+                        "Remove permission " + permission.getName() + " from identity " + identityName);
+            }
+        });
+    }
+
+    private void updatePassword(final String identityName, final PasswordConfiguration passwordData, final User user)
+            throws KuraException {
+
+        final Dictionary<String, Object> properties = user.getProperties();
+
+        final Object currentIsPasswordChangeNeeded = properties.get(IdentityServiceImpl.KURA_NEED_PASSWORD_CHANGE);
+
+        if (passwordData.isPasswordChangeNeeded()) {
+            if (!"true".equals(currentIsPasswordChangeNeeded)) {
+                IdentityServiceImpl.audit(() -> setProperty(properties,
+                        IdentityServiceImpl.KURA_NEED_PASSWORD_CHANGE, "true"),
+                        "Enable password change at next login for identity " + identityName);
+            }
+        } else if (currentIsPasswordChangeNeeded != null) {
+            IdentityServiceImpl.audit(() -> removeProperty(properties, IdentityServiceImpl.KURA_NEED_PASSWORD_CHANGE),
+                    "Disable password change at next login for identity " + identityName);
+        }
+
+        final Dictionary<String, Object> credentials = user.getCredentials();
+        final Optional<char[]> newPassword = passwordData.getNewPassword();
+
+        final Object currentPasswordHash = credentials.get(IdentityServiceImpl.PASSWORD_PROPERTY);
+
+        if (passwordData.isPasswordAuthEnabled() && newPassword.isPresent()) {
+
+            IdentityServiceImpl.audit(() -> setProperty(credentials, IdentityServiceImpl.PASSWORD_PROPERTY,
+                    this.passwordHasher.hash(newPassword.get()).toString()),
+                    "Update Kura password for identity " + identityName);
+
+        } else if (!passwordData.isPasswordAuthEnabled() && currentPasswordHash != null) {
+            IdentityServiceImpl.audit(() -> removeProperty(credentials, IdentityServiceImpl.PASSWORD_PROPERTY),
+                    "Disable Kura password for identity " + identityName);
+        }
+    }
+
+    private void updateAdditionalConfigurations(final String identityName,
+            final AdditionalConfigurations additionalConfigurations) throws KuraException {
+
+        final FailureHandler failureHandler = new FailureHandler(this.logger);
+
+        for (final ComponentConfiguration config : additionalConfigurations.getConfigurations()) {
+            final String pid = config.getPid();
+
+            final Optional<IdentityConfigurationExtension> extension = Optional.ofNullable(this.extensions.get(pid));
+
+            if (!extension.isPresent()) {
+                failureHandler.addError("Configuration extension pid is not registered");
+                continue;
+            }
+
+            try {
+                IdentityServiceImpl.audit(() -> extension.get().updateConfiguration(identityName, config),
+                        "Update configuration for extension " + pid + " for identity " + identityName);
+            } catch (final KuraException e) {
+                failureHandler.addError(e.getMessage());
+            }
+        }
+
+        failureHandler.throwIfFailuresOccurred(KuraErrorCode.CONFIGURATION_ERROR);
+    }
+
+    private void setProperty(final Dictionary<String, Object> properties, final String key, final Object value) {
+        if (!Objects.equals(properties.get(key), value)) {
+            properties.put(key, value);
+        }
+    }
+
+    private void removeProperty(final Dictionary<String, Object> properties, final String key) {
+        if (properties.get(key) != null) {
+            properties.remove(key);
+        }
+    }
+
+    private static final class FailureHandler {
+
+        private final Set<String> errors = new HashSet<>();
+        private final Logger logger;
+
+        private FailureHandler(final Logger logger) {
+            this.logger = logger;
+        }
+
+        private void addError(final String message) {
+            this.errors.add(message);
+            this.logger.error(message);
+        }
+
+        private void throwIfFailuresOccurred(final KuraErrorCode errorCode) throws KuraException {
+            if (!this.errors.isEmpty()) {
+                throw new KuraException(errorCode, this.errors.stream().collect(Collectors.joining("; ")));
+            }
+        }
+    }
+}
